# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Сервис бронирования переговорок",
        "title": "Room Booking Service",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/bookings/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookings"
                ],
                "summary": "Создать бронь на слот (только user).",
                "parameters": [
                    {
                        "description": "Данные для создания бронирования",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_internships-backend_test-backend-barashF_internal_handler_dto_booking.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Бронь создана",
                        "schema": {
                            "$ref": "#/definitions/github_com_internships-backend_test-backend-barashF_internal_handler_dto_booking.BookingCreateResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "$ref": "#/definitions/github_com_internships-backend_test-backend-barashF_internal_handler_dto_booking.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/github_com_internships-backend_test-backend-barashF_internal_handler_dto_booking.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещён (бронирование доступно только роли user)",
                        "schema": {
                            "$ref": "#/definitions/github_com_internships-backend_test-backend-barashF_internal_handler_dto_booking.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Слот не найден",
                        "schema": {
                            "$ref": "#/definitions/github_com_internships-backend_test-backend-barashF_internal_handler_dto_booking.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Слот уже занят",
                        "schema": {
                            "$ref": "#/definitions/github_com_internships-backend_test-backend-barashF_internal_handler_dto_booking.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/github_com_internships-backend_test-backend-barashF_internal_handler_dto_booking.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/bookings/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Доступно только роли admin. Поддерживает пагинацию.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookings"
                ],
                "summary": "Список всех броней с пагинацией (только admin)",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "Номер страницы (начиная с 1). По умолчанию 1.",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "description": "Количество записей на странице. По умолчанию 20, максимум 100.",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список всех броней и метаданные пагинации",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Неверный запрос (некорректные параметры пагинации)",
                        "schema": {
                            "$ref": "#/definitions/github_com_internships-backend_test-backend-barashF_internal_handler_dto_booking.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/github_com_internships-backend_test-backend-barashF_internal_handler_dto_booking.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещён (только admin)",
                        "schema": {
                            "$ref": "#/definitions/github_com_internships-backend_test-backend-barashF_internal_handler_dto_booking.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/github_com_internships-backend_test-backend-barashF_internal_handler_dto_booking.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/bookings/my": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Доступно только роли user. Возвращает брони пользователя.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookings"
                ],
                "summary": "Список броней текущего пользователя (только user)",
                "responses": {
                    "200": {
                        "description": "Список броней текущего пользователя",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Доступ запрещён (только user)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/bookings/{bookingId}/cancel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Доступно только роли user. Пользователь может отменить только свою бронь.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookings"
                ],
                "summary": "Отменить бронь (только своя бронь, только user)",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Идентификатор брони (UUID)",
                        "name": "bookingId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Бронь отменена (или уже была отменена ранее)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Неверный формат ID брони",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Не своя бронь или роль не user",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Бронь не найдена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_internships-backend_test-backend-barashF_internal_handler_dto_booking.BookingCreateResponse": {
            "type": "object",
            "properties": {
                "booking": {
                    "$ref": "#/definitions/github_com_internships-backend_test-backend-barashF_internal_handler_dto_booking.Response"
                }
            }
        },
        "github_com_internships-backend_test-backend-barashF_internal_handler_dto_booking.CreateRequest": {
            "type": "object",
            "properties": {
                "create_conference_link": {
                    "type": "boolean"
                },
                "slot_id": {
                    "type": "string"
                }
            }
        },
        "github_com_internships-backend_test-backend-barashF_internal_handler_dto_booking.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "текст ошибки или причина отказа"
                }
            }
        },
        "github_com_internships-backend_test-backend-barashF_internal_handler_dto_booking.Response": {
            "type": "object",
            "properties": {
                "conference_link": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "slot_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
    }
}